Compilation without warnings
Former-commit-id: 2677c612cd1d81804d11fd923b97d892a8fc53e7

+++ applications/tests/test_fringe_processing/test_fringe_processing_main.cpp
@@ -49,8 +49,8 @@ protected:
 
 TEST_F( FringeProcessingTests, simulPattern)
 {
-    int nx = 311;
-    int ny = 312;
+    size_t nx = 311;
+    size_t ny = 312;
     double noiseLevel = 0.0;
     double freq = 20;
     Matrix1D<int> coefs(10);
